demo/pyproject.toml
[project]
name = "fastgpt-demo"
version = "1.0.0"
description = "FastGPT 文档处理流水线 Demo - Python 实现"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.40",
    "pymupdf>=1.24.0",
    "mammoth>=1.8.0",
    "openpyxl>=3.1.0",
    "markdownify>=0.14.0",
    "beautifulsoup4>=4.12.0",
    "python-pptx>=1.0.0",
    "chardet>=5.0.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastgpt_demo"]
